Allow local versions of the API/UI to be run
This adds `--local-ui` and `--local-api` flags to the `server start`
command, which enabled the local versions of the app to be run,
with the help of Tilt’s `local_resource` function. This requires
environment variables to be set so tilt knows where the repos are,
but this seems to work!