Poll Community API health endpoint for readiness
The Community API is the last thing that comes up when bringing the
stack up. Before, we were just checking the port, which shows as up as
soon as the container comes up. What we’re really interested in is if
the HTTP server is up, so let’s poll the health endpoint instead.