Add a refresh flag
This will allow us to pull all the containers, delete the local
databases, and start from scratch.

To make this a bit more robust, I’m using `yq` to fetch all the
containers defined in the docker-compose file. With this in mind,
I’ve added a Brewfile to automatically install the dependencies
if the user has Homebrew installed.