TEST: multiple domains
We copy the public source hmpps-auth seed[1], remove our entry and readd it with new trusted domains on the end.

When the —local-ui flag is passed we keep the default frontend arround so we can have two frontend services on localhost:3000 and 127.0.0.2:3002 to simulate our test.

To be used in conjunction with this CAS3 frontend test branch: https://github.com/ministryofjustice/hmpps-temporary-accommodation-ui/tree/test-multiple-domain-sign-in

[1] https://github.com/ministryofjustice/hmpps-auth/blob/58d219d53f48e701c6650879589fc802b198cfbd/src/main/resources/db/dev/data/auth/V900_0__clients.sql#L5