Add auto-update functionality to ap-tools command
This provides a convenience functionality to ap-tools so that upstream
changes are propagated to users without manual intervention, minimising
errors due to outdated integrations.

Updates will occur when:
1. The current ap-tools branch is `main` (to avoid trampling over any
   potential changes in-flight)
2. The upstream version of `main` is different to the local version
3. The upstream version can be successfully `git pull`ed

If an update occurs, this will also automatically trigger a container
refresh (as if the `--refresh` flag had been specified).

Automatic updates can be prevented by using the `--no-update` flag.