Mock response from `probation-offender-search`
We need to allow a CAS2 user to search for an offender by prison nubmber
(`nomsNumber`). This adds a mocked response from the
`probation-offender-search` API for searches using Aadland Bertland's
`nomsNumber`. The shape of the data returned should match a prisoner
detail from the probation API, which is a list of `OffenderDetails`.

See their controller tests:

https://github.com/ministryofjustice/probation-offender-search/blob/7093fdf716f969de3b80827675f8389cef01f95c/src/test/kotlin/uk/gov/justice/hmpps/probationsearch/controllers/OffenderSearchControllerTest.kt

And test fixtures:

https://github.com/ministryofjustice/probation-offender-search/blob/a6add4b33fd2aca9928da2cdd4bd276a101dd8b3/src/test/resources/searchdata/james-brown.json