add email for admin
This is to avoid seeing a reminder form to add an email, and ensures our
local setup matches test users set up in dev.

I've added this in the seeding for hmpps-auth because the user's email
is seemingly stored on the HMPPS auth side, which then goes to the
nomis-user-roles API to get the role of the user.