Add all local users to community api staff db
This commit ensures that all of the users loaded by default when running the approved premises API locally also exist in the community API. This ensures calls to the community API get staff endpoint will never 404.

This change also ensures that the probation area id assigned to staff matches the expected values in the approved premises API (i.e. they match the values we seed when preopulated the approved premises API users table)