Add new wiremock mappings for CAS2 E2E tests
The CAS2 frontend wants to write a new E2E test for the application journey where the oasys import fails with a 404[1] (this happens quite often and so we want to cover it).

These new mappings help us replicate what happens when a real oasys record exists but it’s older than 6 months.

We use the wiremock priority configuration to make sure our specific path match takes precedent over the previous wildcard. We give the previous priority 2.

[1] https://github.com/ministryofjustice/hmpps-approved-premises-api/blob/main/src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/controller/cas2/PeopleController.kt#L51