change nomsNumber for our local secondary offender
This was previously added to provide a route for
an offender with no Oasys locally, but the
nomsNumber did not have a corresponding entry in
the Prison API so we were not able to retrieve
prison information for the person, which would
cause our checks on prison codes to fail.

This links it with seeded data in the Prison API
[1]

I'm leaving in the old nomsNumber for now in order
not to break the UI e2e tests for CAS2.

[1] https://github.com/ministryofjustice/prison-api/blob/37b7392bc69960f159de27e72aee19a9ae7e6ad5/src/main/resources/db/migration/data/R__1_15__OFFENDERS.sql#L26