Replace catchall user stubbing in HmppsAuthAndDelius_GetUser/*
We need to be able to log into our local CAS1 development
env with a range of users in order to perform our E2E tests.

This commit replaces the existing "catchall" which returned
the JIMSNOWLDAP details for all requests to

GET ap-and-delius/user/*

with 3 stubs for:

- Jim Snow (`JIMSNOWLDAP`)
- Bernard Beaks (`BERNARD.BEAKS`)
- Approved Premises Test User (`APPROVEDPREMISESTESTUSER`)

NOTE: the pattern matching on the request `urlPathPattern` is exact
so you need to use the Delius usernames in uppercase form.