Make HMPPS-Auth to Delius /user/* stubs case-sensitive
When logging in to HMPPS-Auth the username can be supplied
in any case. With this commit we make our stubs case-insensitive
also so that local development and local E2E tests benefit from
the same behaviour.

[Issue 512][] on the Wiremock repo revealed that the use of "embedded
flags" is supported. This works in the `urlPattern` matcher as
well as for the illustrated `queryParameters`.

[Issue 512][]:
https://github.com/wiremock/wiremock/issues/512#issuecomment-291535632