Use docker volume for postgres
Docker performance when writing to a mapped folder is proving to be poor when used for postgres (e.g. over 1 minute to start from scratch vs 10 seconds if using a docker managed volume).

This coupled with the fact that when using --local-api tilt will only check if postgres is running (not ready) before starting the API is leading to start ups failing

This commit looks to provide stability by:

* Using a docker managed volume for postgres data (as was the case before the previous commit that instead mapped the data to ‘databases/api’
* Update —clear-databases option on stop to also delete docker managed volumes
* Show STDERR messages when starting tilt
* Add a pause between tilt stopping and clearing databases (if required) to ensure that probation-integration has enough time to unlock its H2 database
* Adds misc. notes in the configuration to explain why things are done this way
* Improve the postgres ready check

It’s probably worth looking at using docker managed volumes for all databases for perforamnce reasons instead of mapped folders