Make api config comprehensive
Before this commit we were relying on default config in the api projects application-local.yml to configure various connections details. This commit copies that config into the locally managed API config to give a more comprehensive view of the configuration being used. In time we should remove this duplication from application-local.yml

Furthermore, this commit adds some additional configuration that we’ll need to change/override if testing to upstream dev services for end to end testing