Add auth db for external users
We’re currently having some issues with the hmpps-external-users-api flyway scripts being out of sync with the hmpps-auth flyway scripts. As they share a database this is causing problems on startup. This commit introduces a separate database for hmpps-external-users-api until this problem is resolved. This should be fine because we only read from this database and make use of the seeded test users