Revise API config and document the templates
Add any configuration defined in the API local.yml into the api.template, and document the purpose of these files.

This includes an update to `resolve_secrets.sh` to remove any lines starting with a # when processing a template. This avoids issues when trying to `export` the template values as ENV VARS for npm