Specify postgres platform
There is not a native docker image available for postgis. Where as docker would previously download and use a different platform if there is no local image provided (regardless of platform), this is no longer the case. Therefore, we explicitly specify which platform to use for postgis

i’ve also added `.env.ui` to the .gitignore to avoid accidental checkins when switching from the firebreak ap tools branch